Fetch action items from standalone endpoint with Suspense boundary
